Fix maxLinMotorForce for btSliderConstraint
In btSliderConstraint::getInfo2NonVirtual,
the maxLinMotorForce should be multiplied by
the simulation time step to compute an impulse
that is used to set up the constraints for the
linear motor. Instead of multiplying by timestep,
it was multiplying by fps, so it was effectively
dividing by the timestep. This fixes the error
by dividing by fps.