Limit compilation parallelism
On smaller platforms like the Raspberry Pi, which have a deficit of
physical memory, the default algorithm for determining compilation
parallelism is far too optimistic and usually results in crashing the
platform.

This modification limits the number of parallel jobs to the size of RAM
in Gb. On a mid-size PC (e.g. quad-core with 8+Gb of RAM), this will
make no difference. On a Pi (quad-core with 1Gb of RAM), this will limit
it to 1 job at a time (which is ideal).