Split ConvexInternalShapeData into two structs, one for serialization with double precision, one without
This ends up having quite a few downstream effects, since there are
other shapes that rely on this struct. Those are handled in follow on
commits, and then the DNA is updated at the end of all of them.

The purpose of this work is so that when bullet is compiled with double
precision, serialization is also in double precision, so that
serialization can be deterministic.