Expose collision shape in RayCast callbacks
I'm using a btCompoundShape to combine a bunch of shapes. When I hit a ray I need the specific shapes. Unfortunately, this is currently impossible because btCollisionObject:getShape() doesn't return the individual shape. This can only be retrieved from the btCollisionObjectWrapper or when the result is constructed.

The proposed changes expose the shape directly, allowing me to get the actual shape that was inside of the compound