Fix strict aliasing issues in gim_math.h
The macros GIM_IR(), GIM_SIR(), GIM_AIR(), and GIM_FR() dereference
type-punned pointers, which is UB. The functions gim_inv_sqrt() and
gim_sqrt() use these broken macros.

Introduce two small inline functions copying the bits between the
icompatible float and integer types using memcpy() and rewrite the
macros so they use these functions.

Note that three of the macros returned a reference before the change
and thus could theoretically be used as lvalue; this is no longer
possible. I couldn't find such usage in the code base and it would be
rather odd for 3rd party code to abuse the API like that. However, if
it is required that these macros return lvalues, then a more complicated
approach would be needed with a custom class wrapping the reference
and performing the copy on each access.