Fix strict aliasing issue in b3GpuPgsContactSolver.cpp
csCfg of type b3ConstraintCfg is type-punned to b3SolverBase::ConstraintCfg&
in solveContacts(). The two types are incompatible and this is UB.
But they have almost the same definition, only m_staticIdx is initialized
to a different value in the constructor. Declare b3ConstraintCfg to be
a subclass of b3SolverBase::ConstraintCfg so we can convert between them.