Fix invalid int to pointer conversion in btSoftBody::pointersToIndices()
The code tries to reinterpret an int as a void* with type-punning.
This is not only a strict aliasing violation, but completely
invalid on architectures where ints and pointers have different size.

The leaf btDbvtNode's payload is a union which contains an int dataAsInt
member, so we set that instead of void *data.