Fix incorrect linear velocity of the vehicle chassis
When the vehicle is completely stationary its linear velocity obtained with getLinearVelocity in Y axis is always positive and quite high i.e. 0.15. That's not any oscillation but a steady value of a stationary value. I noticed that getCurrentSpeedKmHour is not equal to getLinearVelocity.length() * 3.6 when the simulation step is finished, it only is equal when m_currentVehicleSpeedKmHour is stored. After some investigation I found that impulses added to the vehicle body by the suspension are not being processed until the next simulation step. This change allows the linear velocity to be reported correctly. It may also reduce input lag by one step. I'm only not sure if updateActions shouldn't be even higher. Being above integrateTransforms is enough to fix the linear velocity issue.